% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGTresults.R
\name{addGTresults}
\alias{addGTresults}
\title{Pull out data from a main GLOBETROTTER results table depending on what event-type is called}
\usage{
addGTresults(pltable, rev_pops)
}
\arguments{
\item{pops}{a vector of populations in "pltable" (the default name for the GT results table)
rev_pops a vector of populations where we want to define the main source as the one that contribute least to the admixture event. This is useful when we want to line up populations to have the the source coming from the same place together on a plot, even if the source from one is 'major' and the other'minor'}
}
\value{
this is a rather unwieldy function that adds to the various results tables defined in the script prior to running this.
}
\description{
This function extracts information on admixture events from a GLOBETROTTER results table and applies them some data frames that can be used for plotting.
}
\examples{
addGTresults(pop)
}
\keyword{Busby_bespoke}

