% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/convertfactorstostrings.R
\name{convert.factors.to.strings.in.dataframe}
\alias{convert.factors.to.strings.in.dataframe}
\title{convert factors in a dataframe to strings}
\usage{
convert.factors.to.strings.in.dataframe(dataframe)
}
\arguments{
\item{dataframe}{The dataframe containing factors that you want to convert}
}
\description{
This function converts all the columns containing factors to strings
}
\examples{
convert.factors.to.strings.in.dataframe(dataframe)
}
\keyword{factors}

