% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/findPopRegion.R
\name{findPopRegion}
\alias{findPopRegion}
\title{this is an internal function that converts a list of population names to regions}
\usage{
findPopRegion(pops, popkey)
}
\arguments{
\item{pops}{a list of names of ethnic groups}

\item{popkey}{a table with at least two columns: Ethnic_Group which is the population that you want to look up and RegionM which is the region that the Ethnic_Group}
}
\description{
This function finds the region that an ethnic group.
}
\examples{
findPopsRegion("ESOMALI", popkey)
}
\keyword{Busby_bespoke}

