% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getGlobetrotter.R
\name{getGlobetrotter}
\alias{getGlobetrotter}
\title{extract the relevant information from globetrotter's output files.}
\usage{
getGlobetrotter(infile, donors = NULL)
}
\arguments{
\item{infile}{the file name of the GLOBETROTTER output file}

\item{donors}{a matrix of donor population copying vectors which are used to generate copying vectors for the admixture sources from the source composition betas. Default to NULL, in which case no source copying vectors are generated}
}
\value{
a two element list: the first element is the full results, and the second are the admixture sources as a matrix (should 'donors' be defined above)
}
\description{
This function extracts information on admixture events from a GLOBETROTTER output file
}
\examples{
getGlobetrotter("FULAInolocal.props.out",donors=NULL)
}
\keyword{Busby_bespoke}

