% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGlobetrotterDates.R
\name{getGlobetrotterDates}
\alias{getGlobetrotterDates}
\title{extract the relevant information from globetrotter's date bootstraps output files.}
\usage{
getGlobetrotterDates(inroot, num_boots = 20, suffix = ".txt")
}
\arguments{
\item{inroot}{the root of the file before the bootstrap number, e.g. DIRECTORY/POP.globetrotter.main.bt}

\item{num_boots}{how many different bootstrap files have been generated? Defaults to 20}

\item{suffix}{everything after the BTSTRAPNUMBER in the file name}
}
\value{
a matrix of dates
}
\description{
This function extracts dates from GLOBETROTTER's date output files. It assumes that the dating output files have the form POP.0.boot.txt, or in the case of a null bootstrap, it can either be POP.0.null.boot.txt, in which case use null=T, or POP.null.0.txt, in which case use null2=T. It also assumed that 20 bootstraps have been performed labelled 0 to 19.
}
\examples{
getGlobetrotterDates(...)
}
\keyword{Busby_bespoke}

