% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLines.R
\name{getLines}
\alias{getLines}
\title{internal function for pulling specific lines out of a GLOBETROTTER output file}
\usage{
getLines(x, y, tmp)
}
\arguments{
\item{x}{line number containing headers}

\item{y}{line number containing values}

\item{tmp}{output file that have been scanned in as a variable}
}
\value{
a two row table with the header and values
}
\description{
This function pulls specified lines out of a GLOBETROTTER output file and is used by getGlobetrotter.R
}
\examples{
getLines(12,13,tmp)
}
\keyword{Busby_bespoke}

