% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getMalder.R
\name{getMalder}
\alias{getMalder}
\title{this is an internal function that generates a date from a time in generations}
\usage{
getMalder(file_name, pop1)
}
\arguments{
\item{file_name}{name of file to pull MALDER results from.}

\item{pop1}{name of population.}
}
\description{
This function takes MALDER output and tabulates
}
\examples{
getMalder(...)
}
\keyword{Busby_bespoke}

