% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopSymbols.R
\name{getPopSymbols}
\alias{getPopSymbols}
\title{find colour, pch, and rim colour for an ethnic group}
\usage{
getPopSymbols(pop_vec, leginfo)
}
\arguments{
\item{pop_vec}{a vector (can be length 1) of ethnic groups}

\item{leginfo}{an object containing data to match to the ethnic group, must have the following columns: "EthnicGroup" "Country" "Region" "Colour" "pch" "rim", where rim defines whether the rim is plotted as the same colour as the symbol}
}
\value{
a dataframe containing the ethnic group, colour, pch, and rim colour for points.
}
\description{
This function finds various plotting parameters for a given ethnic group, as outlined in the leginfo object
}
\examples{
getPopSymbols("FULA",leginfo)
}
\keyword{Busby_bespoke}

