% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDate.R
\name{makeDate}
\alias{makeDate}
\title{this is an internal function that turns an FS type matrix into rows averaged across clusters/groups}
\usage{
makeDate(x, year0 = 1950, gen_length = 29, add_BCE = TRUE)
}
\arguments{
\item{x}{a time in generations}

\item{year0}{the year to start counting back from. Defaults to 1950}

\item{gen_length}{generation length in years. Defaults to 29 years}

\item{add_BCE}{if TRUE converts negative years (i.e. before 0CE) to year in BCE. Defaults to TRUE}
}
\description{
This function converts a date in generations to a date in time.
}
\examples{
makeDate(20, gen_length=30,)
}
\keyword{Busby_bespoke}

