% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/makeGenFromDate.R
\name{makeGenFromDate}
\alias{makeGenFromDate}
\title{this is an internal function that generates time in generations from a date}
\usage{
makeGenFromDate(x, year0 = 1950, gen_length = 29)
}
\arguments{
\item{x}{a time in generations}

\item{year0}{the year to start counting back from. Defaults to 1950}

\item{gen_length}{generation length in years. Defaults to 29 years}
}
\description{
This function converts a date to time in generations.
}
\examples{
makeGenFromDate(20, year0=2000, gen_length=30)
}
\keyword{Busby_bespoke}

