% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/makeTransparent.R
\name{makeTransparent}
\alias{makeTransparent}
\title{makes a colour transparent}
\usage{
makeTransparent(someColor, alpha = 140)
}
\arguments{
\item{someColor}{the colour that you want to make transparent}

\item{alpha}{the amount of transparency. Defaults to 140}
}
\description{
This function makes colors transparent.
}
\examples{
makeTransparent("red", 100)
}
\keyword{colors}

