% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/plotPopPoints.R
\name{plotPopPoints}
\alias{plotPopPoints}
\title{this is an internal function that finds new lat/longs for groups where there are multiple populations from the same country and plots onto a map}
\usage{
plotPopPoints(latlong_file, leginfo_file, poppos_file, pt_cex = 1,
  pt_lwd = 0.5)
}
\arguments{
\item{latlong_file}{a CSV file containing all ethnic groups in the analysis, must have the following columns: "EthnicGroup" "Country" "Lat" "Long"}

\item{leginfo_file}{a file matching the ethnic groups to plotting parameters, must have the following columns: "EthnicGroup" "Country" "Region" "Colour" "pch" "rim", where rim defines whether the rim is plotted as the same colour as the symbol}

\item{poppos_file}{a file outlining for particular countries, whether latitude and/or longitude should be shifted up or down from the original country latitude or longitude}

\item{pt_cex}{the size of the symbols on the map}

\item{pt_lwd}{the width of the outline to the symbols}
}
\value{
plots new points and returns a dataframe with plotting parameters (helpful for legends)
}
\description{
This function re-positions points on a map using three different input files, for use with plotting samples on a map, for example of Africa
}
\examples{
plotPopPoints("data/MalariaGenAdmixturePopulationKeyLatLongs.txt",
              "data/MalariaGenAdmixturePopulationKey.txt",
              "data/MalariaGenAdmixturePopulationKeyMapPositions.txt",
              pt_cex=2,pt_lwd=0.75)
}
\keyword{Busby_bespoke}

