% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/rowsAsMapClusts.R
\name{rowsAsMapClusts}
\alias{rowsAsMapClusts}
\title{this is an internal function that turns an FS type matrix into rows averaged across clusters/groups}
\usage{
rowsAsMapClusts(x, y1, stat = mean)
}
\arguments{
\item{x}{a list of clusters, as output as a MapState in D. Lawson's FinestructureLibrary.R}

\item{y1}{a matrix of copying vectors, where each row is an individual labelled as in the clusters in x}

\item{stat}{should rows be averaged or summed together? Defaults to mean average}
}
\description{
This function converts a an fineSTRUCTURE type matrix into rows averaged across clusters/groups.
}
\examples{
rowsAsMapCluste(mapstatelist, fs_matrix)
}
\keyword{Busby_bespoke}

