% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/splitMalder.R
\name{splitMalder}
\alias{splitMalder}
\title{this is an internal function that splits the results of a MALDER analysis}
\usage{
splitMalder(x)
}
\arguments{
\item{x}{a line from a MALDER output file}
}
\description{
This function splits the results of a MALDER analysis into useful components.
}
\examples{
splitmalder(...)
}
\keyword{Busby_bespoke}

