% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/tidyNames.R
\name{tidyNames}
\alias{tidyNames}
\title{this is an internal function that converts a list of names}
\usage{
tidyNames(x, fula = F, pel = F)
}
\arguments{
\item{x}{a list of names}

\item{fula}{should FULAI and FULAII (and MANDINKAI) be kept as is? Defaults to FALSE}

\item{pel}{Should PELII be kept as is? Defaults to FALSE}
}
\description{
This function converts lists of names for various plotting functions used in various plotting functions.
}
\examples{
tidyNames("ESOMALI", fula=T)
}
\keyword{Busby_bespoke}

