% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvd.R
\name{tvd}
\alias{tvd}
\title{this is an internal function that generates a table of TVD estimates bases on a matrix of copying vectors}
\usage{
tvd(x)
}
\arguments{
\item{x}{a matrix of copying vectors}
}
\value{
returns a table of pairwise TVD for each pair of rows of matrix 'x'
}
\description{
This function generates a table of TVD estimates bases on a matrix of copying vectors.
}
\examples{
tvd(x)
}
\keyword{Busby_bespoke}

