% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2p.R
\name{z2p}
\alias{z2p}
\title{this is an internal function that splits the results of a MALDER analysis}
\usage{
z2p(x)
}
\arguments{
\item{x}{a Z-score}
}
\description{
Converts a Z-score to a p-value.
}
\examples{
z2p(3.4)
}
\keyword{Busby_bespoke}

